requirements.txt
setuptools
django>=2
django-crispy-forms>=1.6.0
django-reversion>=2.0.0
django-formtools>=1.0
django-import-export>=0.5.1
django-ckeditor>=5.7.0
httplib2>=0.9.2
future
six

[Excel]
xlwt
xlsxwriter

[Reversion]
django-reversion>=2.0.0
